+++ Peg.h
@@ -8,9 +8,9 @@ struct Peg: public GameObject {
 public:
     Peg();              
     ~Peg() = default;
-    void init(int16_t x, int16_t y, int8_t hits, int8_t color);
-    int16_t getX();
-    int16_t getY();
+    void init(int32_t x, int32_t y, int8_t hits, int8_t color);
+    int32_t getX();
+    int32_t getY();
     uint16_t getW();
     uint16_t getH();
     const uint16_t* getImage();
